pyproject.toml
[project]
name = "daddy_zeus"
version = "0.1.0"
description = "Telegram bot for weather alerts using Open-Meteo and Telegram Bot API."
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==20.7",
    "aiohttp==3.9.1",
    "schedule==1.2.0",
    "python-decouple>=3.8",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pre-commit>=3.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "B", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 